Fix animation on road closing dropdown triangle.
It looks like an update to the SASS compiler does not process fractions passed as arguments to translateY(). When the fraction is escaped by the SASS compiler into the production CSS, it invalidates the entire CSS declaration, preventing the CSS animation from happening. To workaround this issue, I've introduced a variable that has the fraction, and passed that variable as the argument to translateY().